+++ src/component/admin/policyEmployee/PoliciesonEmployee.component.ts
@@ -1,12 +1,16 @@
 import { Component, OnInit } from "@angular/core";
 import { FormBuilder, FormGroup, Validators } from "@angular/forms";
 import { CompanyDetail } from "src/model/companyDetail.model";
-import {  MessageService } from 'primeng/api';
+import {  ConfirmEventType, ConfirmationService, MessageService } from 'primeng/api';
 import { CompanyDetailService } from "src/service/admin/companyDetail.sevice";
 import { Result } from "src/service/result.service";
 import { Router } from "@angular/router";
 import { PoliciesonEmployeeService } from "src/service/admin/policyEmployee.service";
 import { PoliciesonEmployee } from "src/model/policiesonEmployee.model";
+import { Policy } from "src/model/policy.model";
+import { Hospital } from "src/model/hospitalInfo.model";
+import { HospitalInforService } from "src/service/admin/hospitalInfo.service";
+import { PolicyService } from "src/service/admin/policy.service";
 
 @Component({
     templateUrl: './PoliciesonEmployee.component.html'
@@ -21,13 +25,28 @@ export class    PoliciesonEmployeeAdminComponent implements OnInit {
     count : number = 0;
     tableSize : number=10;
     tableSizes :any = [5,10,15,20];
+    policyonEmployeeDetail:PoliciesonEmployee
+    policy: Policy
+    company: CompanyDetail
+    hospital:Hospital
+    visible:boolean =false
+    btn:boolean = false
     constructor(private formBuilder: FormBuilder,
         private messageService: MessageService,
         private policyEmpService: PoliciesonEmployeeService,
-        private router: Router
+        private router: Router,
+        private hospitalService: HospitalInforService,
+        private companyService: CompanyDetailService,
+        private policyService: PolicyService,
+        private confirmationService: ConfirmationService
         ){}
 
     ngOnInit(): void {
+      if(localStorage.getItem('role')=="Admin"){
+        this.btn = true
+      }else{
+        this.btn =false
+      }
      this.getAll();
     }
 
@@ -51,23 +70,63 @@ export class    PoliciesonEmployeeAdminComponent implements OnInit {
 
   }
     async delete(id: any){
-      await this.policyEmpService.delete(id).then(
-        result =>{
-            this.result = result as Result
-        },  
-        err =>{console.log(err)}
-      )
-      console.log(this.result)
-      if(this.result){
-        await this.messageService.add({severity:"success",summary:"Successful",detail:"Delete Policieson Employee Successful"});
-      }
-      else{
-        await this.messageService.add({severity:"error",summary:"Error",detail:"Delete Policieson Employee Fail"});
-    
-      }
-      await this.getAll()
+      await this.confirmationService.confirm({
+        message: 'Are you sure that you want to delete?',
+        header: 'Confirmation',
+        icon: 'pi pi-exclamation-triangle',
+        accept: () => {
+        
+           this.policyEmpService.delete(id).then(
+            result =>{
+                this.result = result as Result
+                if(this.result){
+                   this.messageService.add({severity:"success",summary:"Successful",detail:"Delete Policieson Employee Successful"});
+                }
+                else{
+                   this.messageService.add({severity:"error",summary:"Error",detail:"Delete Policieson Employee Fail"});
+              
+                }
+                 this.getAll()
+            },  
+            err =>{console.log(err)}
+          )
+        },
+        reject: (type) => {
+          var typeS = type as ConfirmEventType
+            switch (typeS) {
+                case ConfirmEventType.REJECT:
+                    this.messageService.add({ severity: 'error', summary: 'Rejected', detail: 'You have rejected' });
+                    break;
+                case ConfirmEventType.CANCEL:
+                    this.messageService.add({ severity: 'warn', summary: 'Cancelled', detail: 'You have cancelled' });
+                    break;
+            }
+        }
+      });
+     
+      
     }
      update(id:any){
       this.router.navigate(["/admin/update-policieson-employee",{id:id}])
+    }
+    async detail(id:any){
+      this.visible=true;
+     await this.policyEmpService.findById(id).then(
+      res=>{
+         this.policyonEmployeeDetail = res as PoliciesonEmployee
+         this.policyService.findById(this.policyonEmployeeDetail.policyId as number).then(
+          ress=>{this.policy = ress as Policy
+            this.hospitalService.findById(this.policy.medicalid).then(
+              res1=>
+              this.hospital = res1 as Hospital
+            )
+            this.companyService.findById(this.policy.companyId).then(
+              res2=>
+              this.company = res2 as CompanyDetail
+            )
+          }
+         ) 
+      }
+      )
     }
 }
